Created the class Total
This class checks if the user enters a valid quantity through the
setValue(String quantityAsString) method, this method ensures that the
quantity input is clear from characters and is greater than 0, and
returns the double type of the input. Then the method itemTotal()
returns the the number of items * the price of the item.